fixed my mistakes
