+++ src/api/users.ts
@@ -1,3 +1,10 @@
+export interface User {
+  id: number;
+  name: string;
+  username: string;
+  email: string;
+}
+
 export default [
   {
     id: 1,
