+++ src/components/TodoList/TodoList.tsx
@@ -1,37 +1,20 @@
-interface User {
-  id: number;
-  name: string;
-  username: string;
-  email: string;
-}
+import { TodoInfo } from '../TodoInfo';
+import type { Todo as ApiTodo } from '../../api/todos';
+import type { User } from '../../api/users';
 
-interface Todo {
-  id: number;
-  title: string;
-  userId: number;
-  completed: boolean;
+type TodoWithUser = ApiTodo & {
   user: User;
-}
+};
 
 interface Props {
-  todos: Todo[];
+  todos: TodoWithUser[];
 }
 
 export const TodoList = ({ todos }: Props) => {
   return (
     <section className="TodoList">
       {todos.map(todo => (
-        <article
-          key={todo.id}
-          data-id={todo.id}
-          className={`TodoInfo ${todo.completed ? 'TodoInfo--completed' : ''}`}
-        >
-          <h2 className="TodoInfo__title">{todo.title}</h2>
-
-          <a className="UserInfo" href={`mailto:${todo.user.email}`}>
-            {todo.user.name}
-          </a>
-        </article>
+        <TodoInfo key={todo.id} todo={todo} />
       ))}
     </section>
   );
